OSIS-4502 move find_roots to repository
	modified:   base/business/education_groups/general_information.py
	modified:   base/forms/learning_unit/search/borrowed.py
	modified:   base/models/group_element_year.py
	modified:   base/models/learning_unit_year.py
	modified:   base/tests/models/test_group_element_year.py
	modified:   base/views/education_groups/detail.py
	modified:   base/views/learning_unit.py
	modified:   education_group/api/views/learning_unit.py
	modified:   program_management/business/group_element_years/detach.py
	new file:   program_management/ddd/repositories/find_roots.py
	modified:   program_management/views/element_utilization.py
	modified:   program_management/views/prerequisite_read.py
	modified:   program_management/views/prerequisite_update.py
	modified:   base/tests/models/test_group_element_year.py
	new file:   program_management/tests/ddd/repositories/test_find_roots.py
