+++ base/models/learning_unit_year.py
@@ -37,10 +37,10 @@ from django.utils.functional import cached_property
 from django.utils.translation import gettext_lazy as _
 from reversion.admin import VersionAdmin
 
+from program_management.ddd import repositories
 from backoffice.settings.base import LANGUAGE_CODE_EN
 from base.business.learning_container_year import get_learning_container_year_warnings
 from base.models import entity_version
-from base.models import group_element_year
 from base.models.academic_year import compute_max_academic_year_adjournment, AcademicYear
 from base.models.entity_version import get_entity_version_parent_or_itself_from_type
 from base.models.enums import active_status, learning_container_year_types
@@ -512,7 +512,7 @@ class LearningUnitYear(SerializableModel):
         ).exists()
 
     def has_or_is_prerequisite(self, education_group_year):
-        formations = group_element_year.find_roots([education_group_year])[education_group_year.id]
+        formations = repositories.find_roots.find_roots([education_group_year])[education_group_year.id]
         return PrerequisiteItem.objects.filter(
             Q(prerequisite__learning_unit_year=self, prerequisite__education_group_year__in=formations) |
             Q(prerequisite__education_group_year__in=formations, learning_unit=self.learning_unit)
