+++ program_management/views/prerequisite_read.py
@@ -27,6 +27,7 @@
 from django.db.models import Prefetch
 from django.utils.translation import gettext_lazy as _
 
+import program_management.ddd.repositories.find_roots
 from base.business.education_groups import perms
 from base.models import group_element_year
 from base.models.education_group_year import EducationGroupYear
@@ -97,7 +98,7 @@ class LearningUnitPrerequisiteGroup(LearningUnitGenericDetailView):
         context = super().get_context_data()
 
         learning_unit_year = context["learning_unit_year"]
-        formations_id = group_element_year.find_roots([learning_unit_year]).get(
+        formations_id = program_management.ddd.repositories.find_roots.find_roots([learning_unit_year]).get(
             learning_unit_year.id,
             []
         )
